src/requirements.txt
aiohttp==3.13.2
requests==2.32.3
restate-sdk>=0.5.1